Added parser

+++ src/edu/cnm/deepdive/calculator/Calculator.java
@@ -3,5 +3,5 @@ package edu.cnm.deepdive.calculator;
 public interface Calculator {
 
 
-  Integer calculate(String postfixString);
+  Double calculate(String postfixString);
 }
